Change tray icon lookup
Instead of using the current theme's "liferea" icon, we attempt to find
a "feedreader" icon in the theme.

If that fails, GTK_STOCK_EXECUTE is displayed instead.
In the future, an icon will be included with this software that will be
used instead of GTK_STOCK_EXECUTE.

On a side note, the icon code has been moved from main.c to gui.c.